Track deletion of merged constants
This fixes an intermittent segfault in the Diffractor test.
The error very closely depends on the order of events,
but essentially what happens is that a
GlobalVariable gets stored into mergedConstants, but then
deleted in jl_merge_module when the opaque closure module
is merged into its parent. A test case is included, but it
is quite specific to current codegen, so it'll probably stop
being useful the next time we restructure things. It's also
intermittent (though would show up under valgrind/asan).
Hopefully, it'll nevertheless be useful.