Resolve artifact ambiguities using shortest match
This changes the behavior of `select_platform()` to resolve ambiguities
using a two-step process.  Previously, we would simply sort the
resultant triplets and return the last one (so as to provide asemi-
arbitrary stability, and also to prefer higher version numbers over
lower version numbers in tags).  However, with the proliferation of tags
(and especially the new `sanitize=memory` tags) we need a way to exclude
these "extra" tags when our `HostPlatform()` does not have them.

This new matching algorithm excludes candidates from matching with the
platform if there are other candidates with fewer total tags.  This
results in a "simplest match first" behavior, which better represents
the intent of tags in the first place.